Addons - split widget and utils in two scripts
